fix: update golangci-lint installation process
Remove the moving step for golangci-lint after installation. This simplifies the Makefile and reduces potential errors during the installation process. It improves the usability of the lint target by ensuring that the golangci-lint binary is correctly placed in the bin directory without additional commands.

Breaking changes: The golangci-lint binary will now be directly installed in the "bin" directory without being moved.

Signed-off-by: Nick Aguilar <nickaguilarh@gmail.com>