feat: add update command to manage dependencies
Introduce an update command in the Makefile for managing dependencies. This simplifies the process of updating to the latest versions of dependencies. The go.mod file is modified to remove an obsolete replace directive. Changes are made to go.sum to reflect the updated dependencies.

Breaking changes: Removed replace directive for hostmetricsreceiver in go.mod.

Signed-off-by: Nick Aguilar <nickaguilarh@gmail.com>